requirements.txt
Django==5.2.1
djangorestframework==3.15.1
drf-spectacular==0.27.1
python-whois
dnspython
requests
python-dotenv
